+++ tests/PHPStan/Analyser/NodeScopeResolverTest.php
@@ -39,6 +39,7 @@ class NodeScopeResolverTest extends TypeInferenceTestCase
 
 		require_once __DIR__ . '/data/instanceof.php';
 
+		yield from $this->gatherAssertTypes(__DIR__ . '/data/call-user-func-array.php');
 		if (PHP_VERSION_ID >= 80000) {
 			yield from $this->gatherAssertTypes(__DIR__ . '/data/call-user-func-array-php8.php');
 		} else {
